feature(auth): refactor authentication pages - create common stylistic template for login and register page
