+++ client/src/pages/login/components/LoginForm.module.scss
@@ -1,0 +1,33 @@
+@use "../../../styles/text" as t;
+@use "../../../styles/colors" as c;
+@use "../../../styles/mixins" as mixins;
+
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 2.4rem;
+}
+
+.actions {
+  display: grid;
+  gap: 1.2rem;
+}
+
+.account_info {
+  color: darken(c.$col-text-white, 10%);
+  display: flex;
+  font-size: t.$text-sm;
+  gap: 0.8rem;
+}
+
+.link {
+  border-radius: 2px;
+  color: c.$col-primary;
+  font-weight: 500;
+  text-decoration: none;
+  @include mixins.focus(c.$col-primary);
+
+  &:hover {
+    color: lighten(c.$col-primary, 10%);
+  }
+}
