+++ client/src/pages/login/components/LoginHeader.module.scss
@@ -1,0 +1,28 @@
+@use "../../../styles/breakpoints" as b;
+@use "../../../styles/colors" as c;
+
+.header {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  gap: 1.2rem;
+}
+
+.logo {
+  height: 48px;
+  width: 128px;
+
+  @media (min-width: b.$sm) {
+    display: none;
+  }
+}
+
+.title {
+  color: c.$col-text-white;
+  font-size: 2.8rem;
+  font-weight: 600;
+}
+
+.subtitle {
+  color: darken(c.$col-text-white, 10%);
+}
