+++ client/src/pages/register/components/RegisterForm.module.scss
@@ -1,0 +1,28 @@
+@use "../../../styles/colors" as c;
+@use "../../../styles/text" as t;
+@use "../../../styles/mixins" as mixins;
+
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 2.4rem;
+}
+
+.actions {
+  display: grid;
+  gap: 1.2rem;
+}
+
+.link {
+  border-radius: 2px;
+  color: c.$col-primary;
+  font-size: t.$text-sm;
+  font-weight: 500;
+  text-decoration: none;
+  width: fit-content;
+  @include mixins.focus(c.$col-primary);
+
+  &:hover {
+    color: lighten(c.$col-primary, 10%);
+  }
+}
