+++ client/src/pages/register/components/RegisterHeader.module.scss
@@ -1,0 +1,15 @@
+@use "../../../styles/colors" as c;
+
+.header {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  gap: 1.2rem;
+  justify-content: center;
+}
+
+.title {
+  color: c.$col-text-white;
+  font-size: 2.8rem;
+  font-weight: 600;
+}
