+++ client/src/pages/register/components/RegisterHeader.tsx
@@ -1,0 +1,9 @@
+import styles from "./RegisterHeader.module.scss";
+
+export const RegisterHeader = () => {
+  return (
+    <div className={styles.header}>
+      <h2 className={styles.title}>Create an account</h2>
+    </div>
+  );
+};
