+++ client/src/pages/templates/auth.page.module.scss
@@ -1,0 +1,47 @@
+@use "../../styles/colors" as c;
+@use "../../styles/text" as t;
+@use "../../styles/breakpoints" as b;
+@use "../../styles/mixins" as mixins;
+
+.wrapper {
+  background-image: url("../../assets/backgrounds/auth_login_background.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+  display: grid;
+  height: 100%;
+  position: relative;
+
+  @media screen and (min-width: b.$sm) {
+    align-items: center;
+    justify-content: center;
+  }
+}
+
+.logo {
+  display: none;
+  height: 48px;
+  left: 3%;
+  position: absolute;
+  top: 3%;
+  width: 98px;
+
+  @media screen and (min-width: b.$sm) {
+    display: flex;
+  }
+}
+
+.content {
+  background-color: #313338;
+  display: flex;
+  flex-direction: column;
+  gap: 2.8rem;
+  height: 100%;
+  padding: 2.4rem;
+
+  @media screen and (min-width: b.$sm) {
+    border-radius: 4px;
+    height: fit-content;
+    padding: 3.2rem;
+    width: 540px;
+  }
+}
